{"schema":"decodex/commit/1","summary":"Split program intake modules","authority":"manual"}
